Optimization

+++ SharpEngine/Components/SpriteSheetComponent.cs
@@ -80,7 +80,7 @@ public class SpriteSheetComponent : Component
             _currentAnim.Length <= 0 || SpriteSize == new Vec2(0) || !Animations.ContainsKey(_currentAnim)) return;
         
         var texture = Entity.Scene.Window.TextureManager.GetTexture(Sprite);
-        var positionSource = new Vec2(SpriteSize.X * (int)(Animations[_currentAnim][_currentImage] % (texture.Width / SpriteSize.X)), SpriteSize.Y * (Animations[_currentAnim][_currentImage] / (int)(texture.Height / SpriteSize.Y)));
+        var positionSource = new Vec2(SpriteSize.X * (int)(Animations[_currentAnim][_currentImage] % (texture.Width / SpriteSize.X)), SpriteSize.Y * (Animations[_currentAnim][_currentImage] / (texture.Height / SpriteSize.Y)));
         Renderer.RenderTexture(Entity.Scene.Window, texture, tc.Position - CameraManager.Position, new Rect(positionSource, SpriteSize), Color.White, MathHelper.ToRadians(tc.Rotation), SpriteSize / 2, tc.Scale, SpriteEffects.None, 1);
     }
